separated commands from mod
