+++ src/commands/script.rs
@@ -1,27 +1,7 @@
-//! This module defines the commands and their execution logic for the cargo-script CLI tool.
-//!
-//! It includes functionalities to run scripts, initialize the Scripts.toml file, and handle script execution.
-use std::{collections::HashMap, env, fs, io, process::Command, sync::{Arc, Mutex}, time::{Duration, Instant}};
-use clap::{Subcommand, ArgAction};
+use std::{collections::HashMap, env, process::Command, sync::{Arc, Mutex}, time::{Duration, Instant}};
 use serde::Deserialize;
 use emoji::symbols;
 use colored::*;
-
-/// Enum representing the different commands supported by the CLI tool.
-#[derive(Subcommand, Debug)]
-pub enum Commands {
-    #[command(about = "Run a script by name defined in Scripts.toml")]
-    Run {
-        #[arg(value_name = "SCRIPT_NAME", action = ArgAction::Set)]
-        script: String,
-        #[arg(short, long, value_name = "KEY=VALUE", action = ArgAction::Append)]
-        env: Vec<String>,
-    },
-    #[command(about = "Initialize a Scripts.toml file in the current directory")]
-    Init,
-    #[command(about = "Show all script names and descriptions defined in Scripts.toml")]
-    Show,
-}
 
 /// Enum representing a script, which can be either a default command or a detailed script with additional metadata.
 #[derive(Deserialize, Debug)]
@@ -42,36 +22,6 @@ pub enum Script {
 pub struct Scripts {
     pub global_env: Option<HashMap<String, String>>,
     pub scripts: HashMap<String, Script>
-}
-
-/// Show all script names and descriptions in a table format.
-/// 
-/// # Arguments
-///
-/// * `scripts` - A reference to the collection of scripts.
-pub fn show_scripts(scripts: &Scripts) {   
-    let mut max_script_name_len = "Script".len();
-    let mut max_description_len = "Description".len();
-
-    for (name, script) in &scripts.scripts {
-        max_script_name_len = max_script_name_len.max(name.len() + 2);
-        let description = match script {
-            Script::Default(_) => "",
-            Script::Detailed { info, .. } => info.as_deref().unwrap_or(""),
-        };
-        max_description_len = max_description_len.max(description.len() + 2);
-    }
-   
-    println!("{:<width1$} {:<width2$}", "Script".yellow(), "Description".yellow(), width1 = max_script_name_len, width2 = max_description_len);
-    println!("{:<width1$} {:<width2$}", "-".repeat(max_script_name_len).yellow(), "-".repeat(max_description_len).yellow(), width1 = max_script_name_len, width2 = max_description_len);
-
-    for (name, script) in &scripts.scripts {
-        let description = match script {
-            Script::Default(_) => "".to_string(),
-            Script::Detailed { info, .. } => info.clone().unwrap_or_else(|| "".to_string()),
-        };
-        println!("{:<width1$} {:<width2$}", name.green(), description, width1 = max_script_name_len, width2 = max_description_len);
-    }
 }
 
 /// Run a script by name, executing any included scripts in sequence.
@@ -275,31 +225,4 @@ fn execute_command(interpreter: Option<&str>, command: &str) {
             }
         }
     }
-}
-
-/// Initialize a Scripts.toml file in the current directory.
-/// If the file already exists, prompt the user for confirmation to replace it.
-pub fn init_script_file() {
-    let file_path = "Scripts.toml";
-    if fs::metadata(file_path).is_ok() {
-        println!("{}  [ {} ] already exists. Do you want to replace it? ({}/{})", symbols::warning::WARNING.glyph, file_path.yellow(), "y".green(), "n".red());
-        let mut input = String::new();
-        io::stdin().read_line(&mut input).expect("Failed to read input");
-        if input.trim().to_lowercase() != "y" {
-            println!("Operation cancelled.");
-            return;
-        }
-    }
-    let default_content = r#"
-[global_env]
-
-[scripts]
-dev = "cargo run"
-build = { command = "cargo build", env = { RUST_LOG = "info" } }
-release = "cargo build --release"
-test = { command = "cargo test", env = { RUST_LOG = "warn" } }
-doc = "cargo doc --no-deps --open"
-"#;
-    fs::write(file_path, default_content).expect("Failed to write Scripts.toml");
-    println!("{}  [ {} ] has been created.", symbols::other_symbol::CHECK_MARK.glyph, "Scripts.toml".green());
 }
